- segmented renderLocationView - cleaned other functions
